Fix stackblitz, remove scss imports, change fd-fieldset to body-less component, remove brs
