+++ apps/docs/src/app/core/component-docs/input-group/examples/input-group-form-example.component.html
@@ -1,5 +1,5 @@
 <form [formGroup]="customForm">
-    <fieldset fd-fieldset class="flex-form-set">
+    <fieldset fd-fieldset>
         <div fd-form-item>
             <label fd-form-label>Search Input</label>
             <fd-input-group
