+++ apps/docs/src/app/core/component-docs/input-group/examples/input-group-number-example/input-group-number-example.component.html
@@ -1,0 +1,4 @@
+<label fd-form-label>Quantity Spinner</label>
+<fd-input-group-number [disabled]="false" [(ngModel)]="numberValue"></fd-input-group-number>
+<br/>
+<span>numberValue: {{numberValue}}</span>
