+++ apps/docs/src/app/core/component-docs/input-group/examples/input-group-number-example/input-group-number-example.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'fd-input-group-number-example',
+    templateUrl: './input-group-number-example.component.html'
+})
+export class InputGroupNumberExampleComponent {
+    numberValue = '123';
+}
