--- apps/docs/src/app/core/component-docs/input-group/examples/input-group-number-form-example.component.html
@@ -1,7 +1,0 @@
-<form [formGroup]="customForm">
-    <label fd-form-label>Quantity Spinner</label>
-    <fd-input-group-number  formControlName="number"></fd-input-group-number>
-</form>
-<br />
-
-Number Entered: {{ customForm.controls?.number?.value }}
