+++ apps/docs/src/app/core/component-docs/input-group/examples/input-group-search-example/input-group-search-example.component.html
@@ -1,0 +1,11 @@
+<label fd-form-label>Search input</label>
+<fd-input-group
+    [disabled]="false"
+    [(ngModel)]="searchTerm"
+    [glyph]="'decline'"
+    [placeholder]="'Search'"
+    [button]="true"
+    (addOnButtonClicked)="searchTerm = ''">
+</fd-input-group>
+<br/>
+<span>searchTerm: {{searchTerm}}</span>
