+++ apps/docs/src/app/core/component-docs/input-group/examples/input-group-search-example/input-group-search-example.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'fd-input-group-search-example',
+    templateUrl: './input-group-search-example.component.html'
+})
+export class InputGroupSearchExampleComponent {
+
+    searchTerm = 'Search Term';
+
+}
