+++ apps/docs/src/app/core/component-docs/input/examples/input-state-example.component.html
@@ -4,7 +4,6 @@
     </label>
     <input fd-form-control type="text" id="input-51" placeholder="Field placeholder text">
     <fd-form-message>Pellentesque metus lacus commodo eget justo ut rutrum varius nunc</fd-form-message>
-    <br/>
 </div>
 <div fd-form-item>
     <label fd-form-label for="input-52">
@@ -12,7 +11,6 @@
     </label>
     <input fd-form-control type="text" id="input-52" placeholder="Field placeholder text" [state]="'valid'">
     <fd-form-message [type]="'success'">Pellentesque metus lacus commodo eget justo ut rutrum varius nunc</fd-form-message>
-    <br/>
 </div>
 <div fd-form-item>
     <label fd-form-label for="input-53">
@@ -20,7 +18,6 @@
     </label>
     <input fd-form-control type="text" id="input-53" placeholder="Field placeholder text" [state]="'invalid'">
     <fd-form-message [type]="'error'">Pellentesque metus lacus commodo eget justo ut rutrum varius nunc</fd-form-message>
-    <br/>
 </div>
 <div fd-form-item>
     <label fd-form-label for="input-54">
@@ -28,7 +25,6 @@
     </label>
     <input fd-form-control type="text" id="input-54" placeholder="Field placeholder text" [state]="'warning'">
     <fd-form-message [type]="'warning'">Pellentesque metus lacus commodo eget justo ut rutrum varius nunc</fd-form-message>
-    <br/>
 </div>
 <div fd-form-item>
     <label fd-form-label for="input-59">
@@ -36,7 +32,6 @@
     </label>
     <input fd-form-control type="text" id="input-59" placeholder="Field placeholder text" [state]="'information'">
     <fd-form-message [type]="'information'">Pellentesque metus lacus commodo eget justo ut rutrum varius nunc</fd-form-message>
-    <br/>
 </div>
 <div fd-form-item>
     <label fd-form-label for="input-55">
