+++ apps/docs/src/app/core/component-docs/multi-input/examples/multi-input-form-example/multi-input-form-example.component.html
@@ -1,5 +1,5 @@
 <form [formGroup]="customForm">
-    <fieldset fd-fieldset class="flex-form-set">
+    <fieldset fd-fieldset>
         <div fd-form-item>
             <label fd-form-label>Valid Localization Editor</label>
             <fd-multi-input
