+++ apps/docs/src/app/core/component-docs/select-native/examples/select-native-form-group-example.component.html
@@ -1,5 +1,5 @@
 <form [formGroup]="customForm">
-    <fieldset fd-fieldset class="flex-form-set">
+    <fieldset fd-fieldset>
         <div fd-form-item>
             <label fd-form-label for="select-1">Default select:</label>
             <select fd-form-control formControlName="selectControl" id="select-1" name="">
