+++ apps/docs/src/app/core/component-docs/select-native/examples/select-native-state-example.component.html
@@ -8,7 +8,6 @@
         <option value="3">Sed bibendum sapien at posuere interdum</option>
     </select>
     <fd-form-message>Pellentesque metus lacus commodo eget justo ut rutrum varius nunc</fd-form-message>
-    <br/>
 </div>
 <div fd-form-item>
     <label fd-form-label for="select-52">
@@ -31,7 +30,6 @@
     </select>
     <fd-form-message [type]="'error'">Pellentesque metus lacus commodo eget justo ut rutrum varius nunc
     </fd-form-message>
-    <br/>
 </div>
 <div fd-form-item>
     <label fd-form-label for="select-54">
@@ -44,7 +42,6 @@
     </select>
     <fd-form-message [type]="'warning'">Pellentesque metus lacus commodo eget justo ut rutrum varius nunc
     </fd-form-message>
-    <br/>
 </div>
 <div fd-form-item>
     <label fd-form-label for="information-select">
@@ -57,7 +54,6 @@
     </select>
     <fd-form-message [type]="'information'">Pellentesque metus lacus commodo eget justo ut rutrum varius nunc
     </fd-form-message>
-    <br/>
 </div>
 <div fd-form-item>
     <label fd-form-label for="select-55">
