+++ apps/docs/src/app/core/component-docs/select/examples/select-forms/select-forms.component.html
@@ -1,5 +1,5 @@
 <form [formGroup]="customForm">
-    <fieldset fd-fieldset class="flex-form-set">
+    <fieldset fd-fieldset>
         <div fd-form-item>
             <fd-select formControlName="selectControl">
                 <fd-option value="apple">Apple</fd-option>
