+++ apps/docs/src/app/core/component-docs/textarea/examples/textarea-state-example.component.html
@@ -4,7 +4,6 @@
     </label>
     <textarea fd-form-control id="textarea-51" placeholder="Field placeholder text"></textarea>
     <fd-form-message>Pellentesque metus lacus commodo eget justo ut rutrum varius nunc</fd-form-message>
-    <br/>
 </div>
 <div fd-form-item>
     <label fd-form-label for="textarea-52">
@@ -13,7 +12,6 @@
     <textarea fd-form-control id="textarea-52" placeholder="Field placeholder text" [state]="'valid'"></textarea>
     <fd-form-message [type]="'success'">Pellentesque metus lacus commodo eget justo ut rutrum varius nunc
     </fd-form-message>
-    <br/>
 </div>
 <div fd-form-item>
     <label fd-form-label for="textarea-53">
@@ -22,7 +20,6 @@
     <textarea fd-form-control id="textarea-53" placeholder="Field placeholder text" [state]="'invalid'"></textarea>
     <fd-form-message [type]="'error'">Pellentesque metus lacus commodo eget justo ut rutrum varius nunc
     </fd-form-message>
-    <br/>
 </div>
 <div fd-form-item>
     <label fd-form-label for="textarea-54">
@@ -31,7 +28,6 @@
     <textarea fd-form-control id="textarea-54" placeholder="Field placeholder text" [state]="'warning'"></textarea>
     <fd-form-message [type]="'warning'">Pellentesque metus lacus commodo eget justo ut rutrum varius nunc
     </fd-form-message>
-    <br/>
 </div>
 <div fd-form-item>
     <label fd-form-label for="information-textarea">
@@ -40,7 +36,6 @@
     <textarea fd-form-control id="information-textarea" placeholder="Field placeholder text" [state]="'information'"></textarea>
     <fd-form-message [type]="'information'">Pellentesque metus lacus commodo eget justo ut rutrum varius nunc
     </fd-form-message>
-    <br/>
 </div>
 <div fd-form-item>
     <label fd-form-label for="textarea-55">
