+++ libs/core/src/lib/form/fieldset/fieldset.component.spec.ts
@@ -1,8 +1,8 @@
-import { FieldsetDirective } from './fieldset.directive';
+import { FieldsetComponent } from './fieldset.component';
 
 describe('FieldsetComponent', () => {
     it('should create an instance', () => {
-        const component = new FieldsetDirective();
+        const component = new FieldsetComponent();
         expect(component).toBeTruthy();
     });
 });
