+++ libs/core/src/lib/form/fieldset/fieldset.component.ts
@@ -1,0 +1,26 @@
+import { Component, HostBinding, ViewEncapsulation } from '@angular/core';
+
+/**
+ * Used for easily displaying forms with a margin. Not necessary for fundamental forms to be functional.
+ *
+ * ```html
+ * <div fd-fieldset
+ *     <div fd-form-item>
+ *         ...
+ *     </div>
+ * </div>
+ * ```
+ */
+@Component({
+    // tslint:disable-next-line:component-selector
+    selector: '[fd-fieldset]',
+    template: `<ng-content></ng-content>`,
+    styleUrls: ['./fieldset.component.scss'],
+    encapsulation: ViewEncapsulation.None
+})
+export class FieldsetComponent {
+
+    /** @hidden */
+    @HostBinding('class.fd-fieldset')
+    fdFieldClass: boolean = true;
+}
