+++ libs/core/src/lib/form/form-item/form-item.component.scss
@@ -1,8 +1,4 @@
 @import "~fundamental-styles/dist/form-item";
 @import "~fundamental-styles/dist/input";
-@import "~fundamental-styles/dist/checkbox";
-@import "~fundamental-styles/dist/radio";
 @import "~fundamental-styles/dist/textarea";
 @import "~fundamental-styles/dist/form-select";
-@import "~fundamental-styles/dist/fieldset";
-
