+++ libs/core/src/lib/form/public_api.ts
@@ -4,4 +4,4 @@ export * from './form-group/form-group.component';
 export * from './form-item/form-item.component';
 export * from './form-legend/form-legend.directive';
 export * from './form-message/form-message.component';
-export * from './fieldset/fieldset.directive';
+export * from './fieldset/fieldset.component';
